SMGCPA: fix region size for array on string initialization [Merged from klever_fixes revisions 38626, 38718]
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@38761 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/SMGState.java
@@ -280,7 +280,7 @@ public class SMGState implements UnmodifiableSMGState, AbstractQueryableState, G
    * @throws SMGInconsistentException when resulting SMGState is inconsistent
    * and the checks are enabled
    */
-  public Optional<SMGRegion> addAnonymousVariable(int pTypeSize)
+  public Optional<SMGRegion> addAnonymousVariable(long pTypeSize)
       throws SMGInconsistentException {
     if (heap.getStackFrames().isEmpty()) {
       return Optional.empty();
